+++ .eslintrc.js
@@ -1,3 +1,6 @@
 module.exports = {
   extends: "@vue-storefront/eslint-config-integrations",
+  rules: {
+    "class-methods-use-this": "off",
+  },
 };