+++ packages/middleware/src/index.ts
@@ -2,3 +2,4 @@ export * from "./types";
 export * from "./createServer";
 export * from "./apiClientFactory";
 export * from "./terminus";
+export { getLogger } from "./logger";