+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change log
 
+## 3.3.0-rc.2
+
+### Minor Changes
+
+rc.1
+
+## 3.3.0-rc.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - @vue-storefront/logger@1.0.0-rc.3
+
+## 3.3.0-rc.0
+
+### Minor Changes
+
+- **[ADDED]** Logger module
+
 ## 3.2.3
 
 ### Patch Changes